+++ src/runcommands/args.py
@@ -7,8 +7,10 @@ from functools import update_wrapper
 from inspect import Parameter as BaseParameter
 from typing import Mapping, Sequence
 
+from cached_property import cached_property
+
 from .exc import CommandError
-from .util import cached_property, invert_string, is_type
+from .util import invert_string, is_type
 
 
 EMPTY = BaseParameter.empty
@@ -441,6 +443,21 @@ class Arg:
             kwargs.pop("type")
 
         return args, kwargs
+
+    def convert_value(self, value: str):
+        """Convert string value to this arg's type."""
+        if not isinstance(value, str):
+            return value
+        if self.is_bool or self.is_bool_or:
+            if value in ("1", "true"):
+                return True
+            elif value in ("0", "false"):
+                return False
+            if self.is_bool:
+                raise ValueError(f"Bool value must be one of 1, true, 0, or false")
+        converter = self.add_argument_args[1]["type"]
+        value = converter(value)
+        return value
 
     def __str__(self):
         kind = "Positional" if self.is_positional else "Optional"