+++ src/runcommands/util/__init__.py
@@ -3,7 +3,6 @@ import inspect
 from collections import OrderedDict
 
 from .data import Data
-from .decorators import cached_property
 from .enums import Color, StreamOptions
 from .misc import abort, flatten_args, format_if, isatty, load_object, merge_dicts
 from .path import abs_path, asset_path, paths_to_str
@@ -15,7 +14,6 @@ from .string import camel_to_underscore, invert_string
 __all__ = [
     "get_commands_in_namespace",
     "Data",
-    "cached_property",
     "Color",
     "StreamOptions",
     "abort",